pyproject.toml
[tool.poetry]
name = "highrise-bot-sdk"
version = "23.1.0.b0"
description = ""
authors = ["Pocket Worlds Inc <server@high.rs>"]
license = "proprietary"
readme = "README.md"
packages = [{include = "highrise", from = "src"}]


[tool.poetry.dependencies]
python = "^3.11"
aiohttp = ">= 3.8.4"
click = ">= 8.1.3"
cattrs = { git = "https://github.com/python-attrs/cattrs.git", branch = "main" }


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.0.1"


[tool.poetry.scripts]
highrise = "highrise.__main__:run"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
